Create README.md
Flight Delay is inevitable and it plays an important role in both profit and loss of the airlines. An accurate estimation of flight delay is critical for airlines because the  result can be applied to increase  customer satisfaction and income of airline.
Tool use - Excel, SQL, Power BI ,Tableau etc.